requirements.txt
PyYAML~=6.0.3
dotenv~=0.9.9
python-dotenv~=1.2.1
Flask~=3.1.2
flask-cors~=6.0.1
requests~=2.32.5
httpx~=0.28.1
notion-client~=2.7.0
networkx~=3.5
numpy~=2.3.5